Scale pixel coordinates for panels,buttons,windows
Instead of storing the scaled height and width as state store whether
the Window / Panel is using scale independent pixels. If so perform the
adjustment in the draw/hitTest functions.

The goal is to avoid mixing scale information into the save state in
ways that will be hard to deal with.

Some regression here, mainly tooltips are not functioning