UI scale (#3159)
* Allow specifying a UI scale factor

Currently this is quite blurry, and the pick hospital screen is broken.

* Keep fonts crisp with UI scaling

* Add settings window for setting the scale

* Fix the cursor position when not drawn

Scale UIResizable, buttons, and panels

* Add scale factor to fonts

The visual difference with TTF fonts is very clear. Yey HiDPI.

* Apply font scale to confirm dialog

* UI scaling for tip of the day

* Add option to draw scaled sprites

* scale the information window

* scale the jukebox

* Scale pixel coordinates for panels,buttons,windows

Instead of storing the scaled height and width as state store whether
the Window / Panel is using scale independent pixels. If so perform the
adjustment in the draw/hitTest functions.

The goal is to avoid mixing scale information into the save state in
ways that will be hard to deal with.

Some regression here, mainly tooltips are not functioning

* Window positions also use scale coordinates

* Draw tooltips correctly for scale

* Add scaling for file/directory dialogs

* Patient dialog

Include fix for line width

* Scale top menu

* Remove redundant drag handling from new_game dialog

* ui scalling for scrollbars and menu list dialogs

* build_room dialog scale

* scale edit room and place objects

* scale advisor

* scale staff profile page

Add the ability to specify flags when loading a bitmap. This is needed
because linear scaling causes artifacts at the border where the face is
composed of multiple images.

* Add scale to staff dialog

* update handling for scaled dialogs

Fonts and graphics need to be reloaded when the flags
changed.

* ui_scale town map

* ui_scale fax

* ui_scale annual reports

* ui_scale bank_manager

* ui_scale drug_casebook

* ui_scale graphs

* ui_scale for hospital policy

* ui_scale progress report

* ui_scale research policy

* ui_scale staff management

* ui_scale machine dialog

* ui_scale machine menu

* ui_scale furnish corridor

* ui_scale queue

* ui_scale update

* ui_scale staff_raise

* ui_scale map_editor

* ui_scale fix language tooltip position

* Fix size of confirm window

Without this fix the confirm window is fixed size even though visually
it adjusted to the size of the text which caused some strange
behaviours. Most visually it was unable to be dragged to the bottom of
the screen.

* Only include the machine button if it fits for scale

Also remove or add the machine button as the window resolution changes.